Uploaded entire program, 90% functional
First Boot: Downloads a list of words from a prior repository and initialises required files, subsequent running doesn't do this. When in the actual menu, 2 options as of current. Check word checks the preloaded list of words and displays any matches, along with the value. Other option exits with code 0. Admin password: admin123 with currently dysfunctional third option, being to add a word to the dictionary :)